requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
colorama==0.4.4
distlib==0.3.1
Django==3.1.3
django_countries==7.2.1
django-crispy-forms==1.10.0
django-easy-pdf==0.1.1
filelock==3.0.12
gunicorn==20.1.0
isort==5.6.4
WeasyPrint==53.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
phone_field==3.3.1
Pillow==8.0.1
pylint==2.6.0
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
virtualenv==20.1.0
virtualenvwrapper-win==1.2.6
whitenoise==5.2.0
wrapt==1.12.1
